Crate syscalls

Modules

raw

Exposes raw syscalls that simply return a usize instead of a Result.

x86_64

Syscalls for the x86_64 architecture.

Macros

raw_syscall

Performs a raw syscall and returns a usize. Use syscall if you wish to get a Result as a return value.

syscall

Performs a syscall and returns a Result<usize, Errno>.

syscall_args

Structs

Errno

SyscallArgs

The 6 arguments of a syscall, raw untyped version.

SysnoMap

A map of syscalls to a type T.

SysnoMapIter

An iterator over the syscall (number, value) pairs contained in a SysnoMap.

SysnoMapValues

An iterator over the syscall values contained in a SysnoMap.

SysnoSet

A set of syscalls.

SysnoSetIter

An iterator over the syscalls contained in a SysnoSet.

Enums

Sysno

Complete list of Linux syscalls.

Traits

ErrnoSentinel

Functions

syscall⚠

Does a raw syscall.

syscall0⚠

Issues a system call with 0 arguments.

syscall1⚠

Issues a system call with 1 argument.

syscall2⚠

Issues a system call with 2 arguments.

syscall3⚠

Issues a system call with 3 arguments.

syscall4⚠

Issues a system call with 4 arguments.

syscall5⚠

Issues a system call with 5 arguments.

syscall6⚠

Issues a system call with 6 arguments.