syscalls::aarch64

Enum Sysno

#[non_exhaustive]

#[repr(i32)]
pub enum Sysno {

    io_setup = 0,
    io_destroy = 1,
    io_submit = 2,
    io_cancel = 3,
    io_getevents = 4,
    setxattr = 5,
    lsetxattr = 6,
    fsetxattr = 7,
    getxattr = 8,
    lgetxattr = 9,
    fgetxattr = 10,
    listxattr = 11,
    llistxattr = 12,
    flistxattr = 13,
    removexattr = 14,
    lremovexattr = 15,
    fremovexattr = 16,
    getcwd = 17,
    lookup_dcookie = 18,
    eventfd2 = 19,
    epoll_create1 = 20,
    epoll_ctl = 21,
    epoll_pwait = 22,
    dup = 23,
    dup3 = 24,
    fcntl = 25,
    inotify_init1 = 26,
    inotify_add_watch = 27,
    inotify_rm_watch = 28,
    ioctl = 29,
    ioprio_set = 30,
    ioprio_get = 31,
    flock = 32,
    mknodat = 33,
    mkdirat = 34,
    unlinkat = 35,
    symlinkat = 36,
    linkat = 37,
    renameat = 38,
    umount2 = 39,
    mount = 40,
    pivot_root = 41,
    nfsservctl = 42,
    statfs = 43,
    fstatfs = 44,
    truncate = 45,
    ftruncate = 46,
    fallocate = 47,
    faccessat = 48,
    chdir = 49,
    fchdir = 50,
    chroot = 51,
    fchmod = 52,
    fchmodat = 53,
    fchownat = 54,
    fchown = 55,
    openat = 56,
    close = 57,
    vhangup = 58,
    pipe2 = 59,
    quotactl = 60,
    getdents64 = 61,
    lseek = 62,
    read = 63,
    write = 64,
    readv = 65,
    writev = 66,
    pread64 = 67,
    pwrite64 = 68,
    preadv = 69,
    pwritev = 70,
    sendfile = 71,
    pselect6 = 72,
    ppoll = 73,
    signalfd4 = 74,
    vmsplice = 75,
    splice = 76,
    tee = 77,
    readlinkat = 78,
    fstatat = 79,
    fstat = 80,
    sync = 81,
    fsync = 82,
    fdatasync = 83,
    sync_file_range = 84,
    timerfd_create = 85,
    timerfd_settime = 86,
    timerfd_gettime = 87,
    utimensat = 88,
    acct = 89,
    capget = 90,
    capset = 91,
    personality = 92,
    exit = 93,
    exit_group = 94,
    waitid = 95,
    set_tid_address = 96,
    unshare = 97,
    futex = 98,
    set_robust_list = 99,
    get_robust_list = 100,
    nanosleep = 101,
    getitimer = 102,
    setitimer = 103,
    kexec_load = 104,
    init_module = 105,
    delete_module = 106,
    timer_create = 107,
    timer_gettime = 108,
    timer_getoverrun = 109,
    timer_settime = 110,
    timer_delete = 111,
    clock_settime = 112,
    clock_gettime = 113,
    clock_getres = 114,
    clock_nanosleep = 115,
    syslog = 116,
    ptrace = 117,
    sched_setparam = 118,
    sched_setscheduler = 119,
    sched_getscheduler = 120,
    sched_getparam = 121,
    sched_setaffinity = 122,
    sched_getaffinity = 123,
    sched_yield = 124,
    sched_get_priority_max = 125,
    sched_get_priority_min = 126,
    sched_rr_get_interval = 127,
    restart_syscall = 128,
    kill = 129,
    tkill = 130,
    tgkill = 131,
    sigaltstack = 132,
    rt_sigsuspend = 133,
    rt_sigaction = 134,
    rt_sigprocmask = 135,
    rt_sigpending = 136,
    rt_sigtimedwait = 137,
    rt_sigqueueinfo = 138,
    rt_sigreturn = 139,
    setpriority = 140,
    getpriority = 141,
    reboot = 142,
    setregid = 143,
    setgid = 144,
    setreuid = 145,
    setuid = 146,
    setresuid = 147,
    getresuid = 148,
    setresgid = 149,
    getresgid = 150,
    setfsuid = 151,
    setfsgid = 152,
    times = 153,
    setpgid = 154,
    getpgid = 155,
    getsid = 156,
    setsid = 157,
    getgroups = 158,
    setgroups = 159,
    uname = 160,
    sethostname = 161,
    setdomainname = 162,
    getrlimit = 163,
    setrlimit = 164,
    getrusage = 165,
    umask = 166,
    prctl = 167,
    getcpu = 168,
    gettimeofday = 169,
    settimeofday = 170,
    adjtimex = 171,
    getpid = 172,
    getppid = 173,
    getuid = 174,
    geteuid = 175,
    getgid = 176,
    getegid = 177,
    gettid = 178,
    sysinfo = 179,
    mq_open = 180,
    mq_unlink = 181,
    mq_timedsend = 182,
    mq_timedreceive = 183,
    mq_notify = 184,
    mq_getsetattr = 185,
    msgget = 186,
    msgctl = 187,
    msgrcv = 188,
    msgsnd = 189,
    semget = 190,
    semctl = 191,
    semtimedop = 192,
    semop = 193,
    shmget = 194,
    shmctl = 195,
    shmat = 196,
    shmdt = 197,
    socket = 198,
    socketpair = 199,
    bind = 200,
    listen = 201,
    accept = 202,
    connect = 203,
    getsockname = 204,
    getpeername = 205,
    sendto = 206,
    recvfrom = 207,
    setsockopt = 208,
    getsockopt = 209,
    shutdown = 210,
    sendmsg = 211,
    recvmsg = 212,
    readahead = 213,
    brk = 214,
    munmap = 215,
    mremap = 216,
    add_key = 217,
    request_key = 218,
    keyctl = 219,
    clone = 220,
    execve = 221,
    mmap = 222,
    fadvise64 = 223,
    swapon = 224,
    swapoff = 225,
    mprotect = 226,
    msync = 227,
    mlock = 228,
    munlock = 229,
    mlockall = 230,
    munlockall = 231,
    mincore = 232,
    madvise = 233,
    remap_file_pages = 234,
    mbind = 235,
    get_mempolicy = 236,
    set_mempolicy = 237,
    migrate_pages = 238,
    move_pages = 239,
    rt_tgsigqueueinfo = 240,
    perf_event_open = 241,
    accept4 = 242,
    recvmmsg = 243,
    wait4 = 260,
    prlimit64 = 261,
    fanotify_init = 262,
    fanotify_mark = 263,
    name_to_handle_at = 264,
    open_by_handle_at = 265,
    clock_adjtime = 266,
    syncfs = 267,
    setns = 268,
    sendmmsg = 269,
    process_vm_readv = 270,
    process_vm_writev = 271,
    kcmp = 272,
    finit_module = 273,
    sched_setattr = 274,
    sched_getattr = 275,
    renameat2 = 276,
    seccomp = 277,
    getrandom = 278,
    memfd_create = 279,
    bpf = 280,
    execveat = 281,
    userfaultfd = 282,
    membarrier = 283,
    mlock2 = 284,
    copy_file_range = 285,
    preadv2 = 286,
    pwritev2 = 287,
    pkey_mprotect = 288,
    pkey_alloc = 289,
    pkey_free = 290,
    statx = 291,
    io_pgetevents = 292,
    rseq = 293,
    kexec_file_load = 294,
    clock_gettime64 = 403,
    clock_settime64 = 404,
    clock_adjtime64 = 405,
    clock_getres_time64 = 406,
    clock_nanosleep_time64 = 407,
    timer_gettime64 = 408,
    timer_settime64 = 409,
    timerfd_gettime64 = 410,
    timerfd_settime64 = 411,
    utimensat_time64 = 412,
    pselect6_time64 = 413,
    ppoll_time64 = 414,
    io_pgetevents_time64 = 416,
    recvmmsg_time64 = 417,
    mq_timedsend_time64 = 418,
    mq_timedreceive_time64 = 419,
    semtimedop_time64 = 420,
    rt_sigtimedwait_time64 = 421,
    futex_time64 = 422,
    sched_rr_get_interval_time64 = 423,
    pidfd_send_signal = 424,
    io_uring_setup = 425,
    io_uring_enter = 426,
    io_uring_register = 427,
    open_tree = 428,
    move_mount = 429,
    fsopen = 430,
    fsconfig = 431,
    fsmount = 432,
    fspick = 433,
    pidfd_open = 434,
    clone3 = 435,
    close_range = 436,
    openat2 = 437,
    pidfd_getfd = 438,
    faccessat2 = 439,
    process_madvise = 440,
    epoll_pwait2 = 441,
    mount_setattr = 442,
    quotactl_fd = 443,
    landlock_create_ruleset = 444,
    landlock_add_rule = 445,
    landlock_restrict_self = 446,
    memfd_secret = 447,
    process_mrelease = 448,
    futex_waitv = 449,
    set_mempolicy_home_node = 450,
    cachestat = 451,
    fchmodat2 = 452,
    map_shadow_stack = 453,
    futex_wake = 454,
    futex_wait = 455,
    futex_requeue = 456,
    statmount = 457,
    listmount = 458,
    lsm_get_self_attr = 459,
    lsm_set_self_attr = 460,
    lsm_list_modules = 461,
    mseal = 462,

}

Complete list of Linux syscalls.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

io_setup = 0

See io_setup(2) for more info on this syscall.

io_destroy = 1

See io_destroy(2) for more info on this syscall.

io_submit = 2

See io_submit(2) for more info on this syscall.

io_cancel = 3

See io_cancel(2) for more info on this syscall.

io_getevents = 4

See io_getevents(2) for more info on this syscall.

setxattr = 5

See setxattr(2) for more info on this syscall.

lsetxattr = 6

See lsetxattr(2) for more info on this syscall.

fsetxattr = 7

See fsetxattr(2) for more info on this syscall.

getxattr = 8

See getxattr(2) for more info on this syscall.

lgetxattr = 9

See lgetxattr(2) for more info on this syscall.

fgetxattr = 10

See fgetxattr(2) for more info on this syscall.

listxattr = 11

See listxattr(2) for more info on this syscall.

llistxattr = 12

See llistxattr(2) for more info on this syscall.

flistxattr = 13

See flistxattr(2) for more info on this syscall.

removexattr = 14

See removexattr(2) for more info on this syscall.

lremovexattr = 15

See lremovexattr(2) for more info on this syscall.

fremovexattr = 16

See fremovexattr(2) for more info on this syscall.

getcwd = 17

See getcwd(2) for more info on this syscall.

lookup_dcookie = 18

See lookup_dcookie(2) for more info on this syscall.

eventfd2 = 19

See eventfd2(2) for more info on this syscall.

epoll_create1 = 20

See epoll_create1(2) for more info on this syscall.

epoll_ctl = 21

See epoll_ctl(2) for more info on this syscall.

epoll_pwait = 22

See epoll_pwait(2) for more info on this syscall.

dup = 23

See dup(2) for more info on this syscall.

dup3 = 24

See dup3(2) for more info on this syscall.

fcntl = 25

See fcntl(2) for more info on this syscall.

inotify_init1 = 26

See inotify_init1(2) for more info on this syscall.

inotify_add_watch = 27

See inotify_add_watch(2) for more info on this syscall.

inotify_rm_watch = 28

See inotify_rm_watch(2) for more info on this syscall.

ioctl = 29

See ioctl(2) for more info on this syscall.

ioprio_set = 30

See ioprio_set(2) for more info on this syscall.

ioprio_get = 31

See ioprio_get(2) for more info on this syscall.

flock = 32

See flock(2) for more info on this syscall.

mknodat = 33

See mknodat(2) for more info on this syscall.

mkdirat = 34

See mkdirat(2) for more info on this syscall.

unlinkat = 35

See unlinkat(2) for more info on this syscall.

symlinkat = 36

See symlinkat(2) for more info on this syscall.

linkat = 37

See linkat(2) for more info on this syscall.

renameat = 38

See renameat(2) for more info on this syscall.

umount2 = 39

See umount2(2) for more info on this syscall.

mount = 40

See mount(2) for more info on this syscall.

pivot_root = 41

See pivot_root(2) for more info on this syscall.

nfsservctl = 42

See nfsservctl(2) for more info on this syscall.

statfs = 43

See statfs(2) for more info on this syscall.

fstatfs = 44

See fstatfs(2) for more info on this syscall.

truncate = 45

See truncate(2) for more info on this syscall.

ftruncate = 46

See ftruncate(2) for more info on this syscall.

fallocate = 47

See fallocate(2) for more info on this syscall.

faccessat = 48

See faccessat(2) for more info on this syscall.

chdir = 49

See chdir(2) for more info on this syscall.

fchdir = 50

See fchdir(2) for more info on this syscall.

chroot = 51

See chroot(2) for more info on this syscall.

fchmod = 52

See fchmod(2) for more info on this syscall.

fchmodat = 53

See fchmodat(2) for more info on this syscall.

fchownat = 54

See fchownat(2) for more info on this syscall.

fchown = 55

See fchown(2) for more info on this syscall.

openat = 56

See openat(2) for more info on this syscall.

close = 57

See close(2) for more info on this syscall.

vhangup = 58

See vhangup(2) for more info on this syscall.

pipe2 = 59

See pipe2(2) for more info on this syscall.

quotactl = 60

See quotactl(2) for more info on this syscall.

getdents64 = 61

See getdents64(2) for more info on this syscall.

lseek = 62

See lseek(2) for more info on this syscall.

read = 63

See read(2) for more info on this syscall.

write = 64

See write(2) for more info on this syscall.

readv = 65

See readv(2) for more info on this syscall.

writev = 66

See writev(2) for more info on this syscall.

pread64 = 67

See pread64(2) for more info on this syscall.

pwrite64 = 68

See pwrite64(2) for more info on this syscall.

preadv = 69

See preadv(2) for more info on this syscall.

pwritev = 70

See pwritev(2) for more info on this syscall.

sendfile = 71

See sendfile(2) for more info on this syscall.

pselect6 = 72

See pselect6(2) for more info on this syscall.

ppoll = 73

See ppoll(2) for more info on this syscall.

signalfd4 = 74

See signalfd4(2) for more info on this syscall.

vmsplice = 75

See vmsplice(2) for more info on this syscall.

splice = 76

See splice(2) for more info on this syscall.

tee = 77

See tee(2) for more info on this syscall.

readlinkat = 78

See readlinkat(2) for more info on this syscall.

fstatat = 79

See fstatat(2) for more info on this syscall.

fstat = 80

See fstat(2) for more info on this syscall.

sync = 81

See sync(2) for more info on this syscall.

fsync = 82

See fsync(2) for more info on this syscall.

fdatasync = 83

See fdatasync(2) for more info on this syscall.

sync_file_range = 84

See sync_file_range(2) for more info on this syscall.

timerfd_create = 85

See timerfd_create(2) for more info on this syscall.

timerfd_settime = 86

See timerfd_settime(2) for more info on this syscall.

timerfd_gettime = 87

See timerfd_gettime(2) for more info on this syscall.

utimensat = 88

See utimensat(2) for more info on this syscall.

acct = 89

See acct(2) for more info on this syscall.

capget = 90

See capget(2) for more info on this syscall.

capset = 91

See capset(2) for more info on this syscall.

personality = 92

See personality(2) for more info on this syscall.

exit = 93

See exit(2) for more info on this syscall.

exit_group = 94

See exit_group(2) for more info on this syscall.

waitid = 95

See waitid(2) for more info on this syscall.

set_tid_address = 96

See set_tid_address(2) for more info on this syscall.

unshare = 97

See unshare(2) for more info on this syscall.

futex = 98

See futex(2) for more info on this syscall.

set_robust_list = 99

See set_robust_list(2) for more info on this syscall.

get_robust_list = 100

See get_robust_list(2) for more info on this syscall.

nanosleep = 101

See nanosleep(2) for more info on this syscall.

getitimer = 102

See getitimer(2) for more info on this syscall.

setitimer = 103

See setitimer(2) for more info on this syscall.

kexec_load = 104

See kexec_load(2) for more info on this syscall.

init_module = 105

See init_module(2) for more info on this syscall.

delete_module = 106

See delete_module(2) for more info on this syscall.

timer_create = 107

See timer_create(2) for more info on this syscall.

timer_gettime = 108

See timer_gettime(2) for more info on this syscall.

timer_getoverrun = 109

See timer_getoverrun(2) for more info on this syscall.

timer_settime = 110

See timer_settime(2) for more info on this syscall.

timer_delete = 111

See timer_delete(2) for more info on this syscall.

clock_settime = 112

See clock_settime(2) for more info on this syscall.

clock_gettime = 113

See clock_gettime(2) for more info on this syscall.

clock_getres = 114

See clock_getres(2) for more info on this syscall.

clock_nanosleep = 115

See clock_nanosleep(2) for more info on this syscall.

syslog = 116

See syslog(2) for more info on this syscall.

ptrace = 117

See ptrace(2) for more info on this syscall.

sched_setparam = 118

See sched_setparam(2) for more info on this syscall.

sched_setscheduler = 119

See sched_setscheduler(2) for more info on this syscall.

sched_getscheduler = 120

See sched_getscheduler(2) for more info on this syscall.

sched_getparam = 121

See sched_getparam(2) for more info on this syscall.

sched_setaffinity = 122

See sched_setaffinity(2) for more info on this syscall.

sched_getaffinity = 123

See sched_getaffinity(2) for more info on this syscall.

sched_yield = 124

See sched_yield(2) for more info on this syscall.

sched_get_priority_max = 125

See sched_get_priority_max(2) for more info on this syscall.

sched_get_priority_min = 126

See sched_get_priority_min(2) for more info on this syscall.

sched_rr_get_interval = 127

See sched_rr_get_interval(2) for more info on this syscall.

restart_syscall = 128

See restart_syscall(2) for more info on this syscall.

kill = 129

See kill(2) for more info on this syscall.

tkill = 130

See tkill(2) for more info on this syscall.

tgkill = 131

See tgkill(2) for more info on this syscall.

sigaltstack = 132

See sigaltstack(2) for more info on this syscall.

rt_sigsuspend = 133

See rt_sigsuspend(2) for more info on this syscall.

rt_sigaction = 134

See rt_sigaction(2) for more info on this syscall.

rt_sigprocmask = 135

See rt_sigprocmask(2) for more info on this syscall.

rt_sigpending = 136

See rt_sigpending(2) for more info on this syscall.

rt_sigtimedwait = 137

See rt_sigtimedwait(2) for more info on this syscall.

rt_sigqueueinfo = 138

See rt_sigqueueinfo(2) for more info on this syscall.

rt_sigreturn = 139

See rt_sigreturn(2) for more info on this syscall.

setpriority = 140

See setpriority(2) for more info on this syscall.

getpriority = 141

See getpriority(2) for more info on this syscall.

reboot = 142

See reboot(2) for more info on this syscall.

setregid = 143

See setregid(2) for more info on this syscall.

setgid = 144

See setgid(2) for more info on this syscall.

setreuid = 145

See setreuid(2) for more info on this syscall.

setuid = 146

See setuid(2) for more info on this syscall.

setresuid = 147

See setresuid(2) for more info on this syscall.

getresuid = 148

See getresuid(2) for more info on this syscall.

setresgid = 149

See setresgid(2) for more info on this syscall.

getresgid = 150

See getresgid(2) for more info on this syscall.

setfsuid = 151

See setfsuid(2) for more info on this syscall.

setfsgid = 152

See setfsgid(2) for more info on this syscall.

times = 153

See times(2) for more info on this syscall.

setpgid = 154

See setpgid(2) for more info on this syscall.

getpgid = 155

See getpgid(2) for more info on this syscall.

getsid = 156

See getsid(2) for more info on this syscall.

setsid = 157

See setsid(2) for more info on this syscall.

getgroups = 158

See getgroups(2) for more info on this syscall.

setgroups = 159

See setgroups(2) for more info on this syscall.

uname = 160

See uname(2) for more info on this syscall.

sethostname = 161

See sethostname(2) for more info on this syscall.

setdomainname = 162

See setdomainname(2) for more info on this syscall.

getrlimit = 163

See getrlimit(2) for more info on this syscall.

setrlimit = 164

See setrlimit(2) for more info on this syscall.

getrusage = 165

See getrusage(2) for more info on this syscall.

umask = 166

See umask(2) for more info on this syscall.

prctl = 167

See prctl(2) for more info on this syscall.

getcpu = 168

See getcpu(2) for more info on this syscall.

gettimeofday = 169

See gettimeofday(2) for more info on this syscall.

settimeofday = 170

See settimeofday(2) for more info on this syscall.

adjtimex = 171

See adjtimex(2) for more info on this syscall.

getpid = 172

See getpid(2) for more info on this syscall.

getppid = 173

See getppid(2) for more info on this syscall.

getuid = 174

See getuid(2) for more info on this syscall.

geteuid = 175

See geteuid(2) for more info on this syscall.

getgid = 176

See getgid(2) for more info on this syscall.

getegid = 177

See getegid(2) for more info on this syscall.

gettid = 178

See gettid(2) for more info on this syscall.

sysinfo = 179

See sysinfo(2) for more info on this syscall.

mq_open = 180

See mq_open(2) for more info on this syscall.

mq_unlink = 181

See mq_unlink(2) for more info on this syscall.

mq_timedsend = 182

See mq_timedsend(2) for more info on this syscall.

mq_timedreceive = 183

See mq_timedreceive(2) for more info on this syscall.

mq_notify = 184

See mq_notify(2) for more info on this syscall.

mq_getsetattr = 185

See mq_getsetattr(2) for more info on this syscall.

msgget = 186

See msgget(2) for more info on this syscall.

msgctl = 187

See msgctl(2) for more info on this syscall.

msgrcv = 188

See msgrcv(2) for more info on this syscall.

msgsnd = 189

See msgsnd(2) for more info on this syscall.

semget = 190

See semget(2) for more info on this syscall.

semctl = 191

See semctl(2) for more info on this syscall.

semtimedop = 192

See semtimedop(2) for more info on this syscall.

semop = 193

See semop(2) for more info on this syscall.

shmget = 194

See shmget(2) for more info on this syscall.

shmctl = 195

See shmctl(2) for more info on this syscall.

shmat = 196

See shmat(2) for more info on this syscall.

shmdt = 197

See shmdt(2) for more info on this syscall.

socket = 198

See socket(2) for more info on this syscall.

socketpair = 199

See socketpair(2) for more info on this syscall.

bind = 200

See bind(2) for more info on this syscall.

listen = 201

See listen(2) for more info on this syscall.

accept = 202

See accept(2) for more info on this syscall.

connect = 203

See connect(2) for more info on this syscall.

getsockname = 204

See getsockname(2) for more info on this syscall.

getpeername = 205

See getpeername(2) for more info on this syscall.

sendto = 206

See sendto(2) for more info on this syscall.

recvfrom = 207

See recvfrom(2) for more info on this syscall.

setsockopt = 208

See setsockopt(2) for more info on this syscall.

getsockopt = 209

See getsockopt(2) for more info on this syscall.

shutdown = 210

See shutdown(2) for more info on this syscall.

sendmsg = 211

See sendmsg(2) for more info on this syscall.

recvmsg = 212

See recvmsg(2) for more info on this syscall.

readahead = 213

See readahead(2) for more info on this syscall.

brk = 214

See brk(2) for more info on this syscall.

munmap = 215

See munmap(2) for more info on this syscall.

mremap = 216

See mremap(2) for more info on this syscall.

add_key = 217

See add_key(2) for more info on this syscall.

request_key = 218

See request_key(2) for more info on this syscall.

keyctl = 219

See keyctl(2) for more info on this syscall.

clone = 220

See clone(2) for more info on this syscall.

execve = 221

See execve(2) for more info on this syscall.

mmap = 222

See mmap(2) for more info on this syscall.

fadvise64 = 223

See fadvise64(2) for more info on this syscall.

swapon = 224

See swapon(2) for more info on this syscall.

swapoff = 225

See swapoff(2) for more info on this syscall.

mprotect = 226

See mprotect(2) for more info on this syscall.

msync = 227

See msync(2) for more info on this syscall.

mlock = 228

See mlock(2) for more info on this syscall.

munlock = 229

See munlock(2) for more info on this syscall.

mlockall = 230

See mlockall(2) for more info on this syscall.

munlockall = 231

See munlockall(2) for more info on this syscall.

mincore = 232

See mincore(2) for more info on this syscall.

madvise = 233

See madvise(2) for more info on this syscall.

remap_file_pages = 234

See remap_file_pages(2) for more info on this syscall.

mbind = 235

See mbind(2) for more info on this syscall.

get_mempolicy = 236

See get_mempolicy(2) for more info on this syscall.

set_mempolicy = 237

See set_mempolicy(2) for more info on this syscall.

migrate_pages = 238

See migrate_pages(2) for more info on this syscall.

move_pages = 239

See move_pages(2) for more info on this syscall.

rt_tgsigqueueinfo = 240

See rt_tgsigqueueinfo(2) for more info on this syscall.

perf_event_open = 241

See perf_event_open(2) for more info on this syscall.

accept4 = 242

See accept4(2) for more info on this syscall.

recvmmsg = 243

See recvmmsg(2) for more info on this syscall.

wait4 = 260

See wait4(2) for more info on this syscall.

prlimit64 = 261

See prlimit64(2) for more info on this syscall.

fanotify_init = 262

See fanotify_init(2) for more info on this syscall.

fanotify_mark = 263

See fanotify_mark(2) for more info on this syscall.

name_to_handle_at = 264

See name_to_handle_at(2) for more info on this syscall.

open_by_handle_at = 265

See open_by_handle_at(2) for more info on this syscall.

clock_adjtime = 266

See clock_adjtime(2) for more info on this syscall.

syncfs = 267

See syncfs(2) for more info on this syscall.

setns = 268

See setns(2) for more info on this syscall.

sendmmsg = 269

See sendmmsg(2) for more info on this syscall.

process_vm_readv = 270

See process_vm_readv(2) for more info on this syscall.

process_vm_writev = 271

See process_vm_writev(2) for more info on this syscall.

kcmp = 272

See kcmp(2) for more info on this syscall.

finit_module = 273

See finit_module(2) for more info on this syscall.

sched_setattr = 274

See sched_setattr(2) for more info on this syscall.

sched_getattr = 275

See sched_getattr(2) for more info on this syscall.

renameat2 = 276

See renameat2(2) for more info on this syscall.

seccomp = 277

See seccomp(2) for more info on this syscall.

getrandom = 278

See getrandom(2) for more info on this syscall.

memfd_create = 279

See memfd_create(2) for more info on this syscall.

bpf = 280

See bpf(2) for more info on this syscall.

execveat = 281

See execveat(2) for more info on this syscall.

userfaultfd = 282

See userfaultfd(2) for more info on this syscall.

membarrier = 283

See membarrier(2) for more info on this syscall.

mlock2 = 284

See mlock2(2) for more info on this syscall.

copy_file_range = 285

See copy_file_range(2) for more info on this syscall.

preadv2 = 286

See preadv2(2) for more info on this syscall.

pwritev2 = 287

See pwritev2(2) for more info on this syscall.

pkey_mprotect = 288

See pkey_mprotect(2) for more info on this syscall.

pkey_alloc = 289

See pkey_alloc(2) for more info on this syscall.

pkey_free = 290

See pkey_free(2) for more info on this syscall.

statx = 291

See statx(2) for more info on this syscall.

io_pgetevents = 292

See io_pgetevents(2) for more info on this syscall.

rseq = 293

See rseq(2) for more info on this syscall.

kexec_file_load = 294

See kexec_file_load(2) for more info on this syscall.

clock_gettime64 = 403

See clock_gettime64(2) for more info on this syscall.

clock_settime64 = 404

See clock_settime64(2) for more info on this syscall.

clock_adjtime64 = 405

See clock_adjtime64(2) for more info on this syscall.

clock_getres_time64 = 406

See clock_getres_time64(2) for more info on this syscall.

clock_nanosleep_time64 = 407

See clock_nanosleep_time64(2) for more info on this syscall.

timer_gettime64 = 408

See timer_gettime64(2) for more info on this syscall.

timer_settime64 = 409

See timer_settime64(2) for more info on this syscall.

timerfd_gettime64 = 410

See timerfd_gettime64(2) for more info on this syscall.

timerfd_settime64 = 411

See timerfd_settime64(2) for more info on this syscall.

utimensat_time64 = 412

See utimensat_time64(2) for more info on this syscall.

pselect6_time64 = 413

See pselect6_time64(2) for more info on this syscall.

ppoll_time64 = 414

See ppoll_time64(2) for more info on this syscall.

io_pgetevents_time64 = 416

See io_pgetevents_time64(2) for more info on this syscall.

recvmmsg_time64 = 417

See recvmmsg_time64(2) for more info on this syscall.

mq_timedsend_time64 = 418

See mq_timedsend_time64(2) for more info on this syscall.

mq_timedreceive_time64 = 419

See mq_timedreceive_time64(2) for more info on this syscall.

semtimedop_time64 = 420

See semtimedop_time64(2) for more info on this syscall.

rt_sigtimedwait_time64 = 421

See rt_sigtimedwait_time64(2) for more info on this syscall.

futex_time64 = 422

See futex_time64(2) for more info on this syscall.

sched_rr_get_interval_time64 = 423

See sched_rr_get_interval_time64(2) for more info on this syscall.

pidfd_send_signal = 424

See pidfd_send_signal(2) for more info on this syscall.

io_uring_setup = 425

See io_uring_setup(2) for more info on this syscall.

io_uring_enter = 426

See io_uring_enter(2) for more info on this syscall.

io_uring_register = 427

See io_uring_register(2) for more info on this syscall.

open_tree = 428

See open_tree(2) for more info on this syscall.

move_mount = 429

See move_mount(2) for more info on this syscall.

fsopen = 430

See fsopen(2) for more info on this syscall.

fsconfig = 431

See fsconfig(2) for more info on this syscall.

fsmount = 432

See fsmount(2) for more info on this syscall.

fspick = 433

See fspick(2) for more info on this syscall.

pidfd_open = 434

See pidfd_open(2) for more info on this syscall.

clone3 = 435

See clone3(2) for more info on this syscall.

close_range = 436

See close_range(2) for more info on this syscall.

openat2 = 437

See openat2(2) for more info on this syscall.

pidfd_getfd = 438

See pidfd_getfd(2) for more info on this syscall.

faccessat2 = 439

See faccessat2(2) for more info on this syscall.

process_madvise = 440

See process_madvise(2) for more info on this syscall.

epoll_pwait2 = 441

See epoll_pwait2(2) for more info on this syscall.

mount_setattr = 442

See mount_setattr(2) for more info on this syscall.

quotactl_fd = 443

See quotactl_fd(2) for more info on this syscall.

landlock_create_ruleset = 444

See landlock_create_ruleset(2) for more info on this syscall.

landlock_add_rule = 445

See landlock_add_rule(2) for more info on this syscall.

landlock_restrict_self = 446

See landlock_restrict_self(2) for more info on this syscall.

memfd_secret = 447

See memfd_secret(2) for more info on this syscall.

process_mrelease = 448

See process_mrelease(2) for more info on this syscall.

futex_waitv = 449

See futex_waitv(2) for more info on this syscall.

set_mempolicy_home_node = 450

See set_mempolicy_home_node(2) for more info on this syscall.

cachestat = 451

See cachestat(2) for more info on this syscall.

fchmodat2 = 452

See fchmodat2(2) for more info on this syscall.

map_shadow_stack = 453

See map_shadow_stack(2) for more info on this syscall.

futex_wake = 454

See futex_wake(2) for more info on this syscall.

futex_wait = 455

See futex_wait(2) for more info on this syscall.

futex_requeue = 456

See futex_requeue(2) for more info on this syscall.

statmount = 457

See statmount(2) for more info on this syscall.

listmount = 458

See listmount(2) for more info on this syscall.

lsm_get_self_attr = 459

See lsm_get_self_attr(2) for more info on this syscall.

lsm_set_self_attr = 460

See lsm_set_self_attr(2) for more info on this syscall.

lsm_list_modules = 461

See lsm_list_modules(2) for more info on this syscall.

mseal = 462

See mseal(2) for more info on this syscall.

Implementations

impl Sysno

pub const fn new(id: usize) -> Option<Self>

Constructs a new syscall from the given ID. If the ID does not represent a valid syscall, returns None.

pub const fn name(&self) -> &'static str

Returns the name of the syscall.

pub const fn next(&self) -> Option<Self>

Returns the next syscall in the table. Returns None if this is the last syscall.

pub const fn first() -> Self

Returns the first syscall in the table.

pub const fn last() -> Self

Returns the last syscall in the table.

pub const fn id(&self) -> i32

Returns the syscall number.

pub const fn len() -> usize

👎Deprecated: Sysno::len() is misleading. Use Sysno::table_size() instead.

Returns the length of the syscall table, including any gaps.

pub const fn count() -> usize

Returns the total number of valid syscalls.

pub const fn table_size() -> usize

Returns the length of the syscall table, including any gaps. This is not the same thing as the total number of syscalls.

pub fn iter() -> impl Iterator<Item = Self>

Returns an iterator that iterates over all possible syscalls.