Enum syscalls::Sysno

#[repr(i32)]
#[non_exhaustive]
pub enum Sysno {

    read,
    write,
    open,
    close,
    stat,
    fstat,
    lstat,
    poll,
    lseek,
    mmap,
    mprotect,
    munmap,
    brk,
    rt_sigaction,
    rt_sigprocmask,
    rt_sigreturn,
    ioctl,
    pread64,
    pwrite64,
    readv,
    writev,
    access,
    pipe,
    select,
    sched_yield,
    mremap,
    msync,
    mincore,
    madvise,
    shmget,
    shmat,
    shmctl,
    dup,
    dup2,
    pause,
    nanosleep,
    getitimer,
    alarm,
    setitimer,
    getpid,
    sendfile,
    socket,
    connect,
    accept,
    sendto,
    recvfrom,
    sendmsg,
    recvmsg,
    shutdown,
    bind,
    listen,
    getsockname,
    getpeername,
    socketpair,
    setsockopt,
    getsockopt,
    clone,
    fork,
    vfork,
    execve,
    exit,
    wait4,
    kill,
    uname,
    semget,
    semop,
    semctl,
    shmdt,
    msgget,
    msgsnd,
    msgrcv,
    msgctl,
    fcntl,
    flock,
    fsync,
    fdatasync,
    truncate,
    ftruncate,
    getdents,
    getcwd,
    chdir,
    fchdir,
    rename,
    mkdir,
    rmdir,
    creat,
    link,
    unlink,
    symlink,
    readlink,
    chmod,
    fchmod,
    chown,
    fchown,
    lchown,
    umask,
    gettimeofday,
    getrlimit,
    getrusage,
    sysinfo,
    times,
    ptrace,
    getuid,
    syslog,
    getgid,
    setuid,
    setgid,
    geteuid,
    getegid,
    setpgid,
    getppid,
    getpgrp,
    setsid,
    setreuid,
    setregid,
    getgroups,
    setgroups,
    setresuid,
    getresuid,
    setresgid,
    getresgid,
    getpgid,
    setfsuid,
    setfsgid,
    getsid,
    capget,
    capset,
    rt_sigpending,
    rt_sigtimedwait,
    rt_sigqueueinfo,
    rt_sigsuspend,
    sigaltstack,
    utime,
    mknod,
    uselib,
    personality,
    ustat,
    statfs,
    fstatfs,
    sysfs,
    getpriority,
    setpriority,
    sched_setparam,
    sched_getparam,
    sched_setscheduler,
    sched_getscheduler,
    sched_get_priority_max,
    sched_get_priority_min,
    sched_rr_get_interval,
    mlock,
    munlock,
    mlockall,
    munlockall,
    vhangup,
    modify_ldt,
    pivot_root,
    _sysctl,
    prctl,
    arch_prctl,
    adjtimex,
    setrlimit,
    chroot,
    sync,
    acct,
    settimeofday,
    mount,
    umount2,
    swapon,
    swapoff,
    reboot,
    sethostname,
    setdomainname,
    iopl,
    ioperm,
    create_module,
    init_module,
    delete_module,
    get_kernel_syms,
    query_module,
    quotactl,
    nfsservctl,
    getpmsg,
    putpmsg,
    afs_syscall,
    tuxcall,
    security,
    gettid,
    readahead,
    setxattr,
    lsetxattr,
    fsetxattr,
    getxattr,
    lgetxattr,
    fgetxattr,
    listxattr,
    llistxattr,
    flistxattr,
    removexattr,
    lremovexattr,
    fremovexattr,
    tkill,
    time,
    futex,
    sched_setaffinity,
    sched_getaffinity,
    set_thread_area,
    io_setup,
    io_destroy,
    io_getevents,
    io_submit,
    io_cancel,
    get_thread_area,
    lookup_dcookie,
    epoll_create,
    epoll_ctl_old,
    epoll_wait_old,
    remap_file_pages,
    getdents64,
    set_tid_address,
    restart_syscall,
    semtimedop,
    fadvise64,
    timer_create,
    timer_settime,
    timer_gettime,
    timer_getoverrun,
    timer_delete,
    clock_settime,
    clock_gettime,
    clock_getres,
    clock_nanosleep,
    exit_group,
    epoll_wait,
    epoll_ctl,
    tgkill,
    utimes,
    vserver,
    mbind,
    set_mempolicy,
    get_mempolicy,
    mq_open,
    mq_unlink,
    mq_timedsend,
    mq_timedreceive,
    mq_notify,
    mq_getsetattr,
    kexec_load,
    waitid,
    add_key,
    request_key,
    keyctl,
    ioprio_set,
    ioprio_get,
    inotify_init,
    inotify_add_watch,
    inotify_rm_watch,
    migrate_pages,
    openat,
    mkdirat,
    mknodat,
    fchownat,
    futimesat,
    newfstatat,
    unlinkat,
    renameat,
    linkat,
    symlinkat,
    readlinkat,
    fchmodat,
    faccessat,
    pselect6,
    ppoll,
    unshare,
    set_robust_list,
    get_robust_list,
    splice,
    tee,
    sync_file_range,
    vmsplice,
    move_pages,
    utimensat,
    epoll_pwait,
    signalfd,
    timerfd_create,
    eventfd,
    fallocate,
    timerfd_settime,
    timerfd_gettime,
    accept4,
    signalfd4,
    eventfd2,
    epoll_create1,
    dup3,
    pipe2,
    inotify_init1,
    preadv,
    pwritev,
    rt_tgsigqueueinfo,
    perf_event_open,
    recvmmsg,
    fanotify_init,
    fanotify_mark,
    prlimit64,
    name_to_handle_at,
    open_by_handle_at,
    clock_adjtime,
    syncfs,
    sendmmsg,
    setns,
    getcpu,
    process_vm_readv,
    process_vm_writev,
    kcmp,
    finit_module,
    sched_setattr,
    sched_getattr,
    renameat2,
    seccomp,
    getrandom,
    memfd_create,
    kexec_file_load,
    bpf,
    execveat,
    userfaultfd,
    membarrier,
    mlock2,
    copy_file_range,
    preadv2,
    pwritev2,
    pkey_mprotect,
    pkey_alloc,
    pkey_free,
    statx,
    io_pgetevents,
    rseq,
    pidfd_send_signal,
    io_uring_setup,
    io_uring_enter,
    io_uring_register,
    open_tree,
    move_mount,
    fsopen,
    fsconfig,
    fsmount,
    fspick,
    pidfd_open,
    clone3,
    close_range,
    openat2,
    pidfd_getfd,
    faccessat2,
    process_madvise,
    epoll_pwait2,
    mount_setattr,
    quotactl_fd,
    landlock_create_ruleset,
    landlock_add_rule,
    landlock_restrict_self,
    memfd_secret,
    process_mrelease,
    futex_waitv,
    set_mempolicy_home_node,

}

Complete list of Linux syscalls.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

read

See read(2) for more info on this syscall.

write

See write(2) for more info on this syscall.

open

See open(2) for more info on this syscall.

close

See close(2) for more info on this syscall.

stat

See stat(2) for more info on this syscall.

fstat

See fstat(2) for more info on this syscall.

lstat

See lstat(2) for more info on this syscall.

poll

See poll(2) for more info on this syscall.

lseek

See lseek(2) for more info on this syscall.

mmap

See mmap(2) for more info on this syscall.

mprotect

See mprotect(2) for more info on this syscall.

munmap

See munmap(2) for more info on this syscall.

brk

See brk(2) for more info on this syscall.

rt_sigaction

See rt_sigaction(2) for more info on this syscall.

rt_sigprocmask

See rt_sigprocmask(2) for more info on this syscall.

rt_sigreturn

See rt_sigreturn(2) for more info on this syscall.

ioctl

See ioctl(2) for more info on this syscall.

pread64

See pread64(2) for more info on this syscall.

pwrite64

See pwrite64(2) for more info on this syscall.

readv

See readv(2) for more info on this syscall.

writev

See writev(2) for more info on this syscall.

access

See access(2) for more info on this syscall.

pipe

See pipe(2) for more info on this syscall.

select

See select(2) for more info on this syscall.

sched_yield

See sched_yield(2) for more info on this syscall.

mremap

See mremap(2) for more info on this syscall.

msync

See msync(2) for more info on this syscall.

mincore

See mincore(2) for more info on this syscall.

madvise

See madvise(2) for more info on this syscall.

shmget

See shmget(2) for more info on this syscall.

shmat

See shmat(2) for more info on this syscall.

shmctl

See shmctl(2) for more info on this syscall.

dup

See dup(2) for more info on this syscall.

dup2

See dup2(2) for more info on this syscall.

pause

See pause(2) for more info on this syscall.

nanosleep

See nanosleep(2) for more info on this syscall.

getitimer

See getitimer(2) for more info on this syscall.

alarm

See alarm(2) for more info on this syscall.

setitimer

See setitimer(2) for more info on this syscall.

getpid

See getpid(2) for more info on this syscall.

sendfile

See sendfile(2) for more info on this syscall.

socket

See socket(2) for more info on this syscall.

connect

See connect(2) for more info on this syscall.

accept

See accept(2) for more info on this syscall.

sendto

See sendto(2) for more info on this syscall.

recvfrom

See recvfrom(2) for more info on this syscall.

sendmsg

See sendmsg(2) for more info on this syscall.

recvmsg

See recvmsg(2) for more info on this syscall.

shutdown

See shutdown(2) for more info on this syscall.

bind

See bind(2) for more info on this syscall.

listen

See listen(2) for more info on this syscall.

getsockname

See getsockname(2) for more info on this syscall.

getpeername

See getpeername(2) for more info on this syscall.

socketpair

See socketpair(2) for more info on this syscall.

setsockopt

See setsockopt(2) for more info on this syscall.

getsockopt

See getsockopt(2) for more info on this syscall.

clone

See clone(2) for more info on this syscall.

fork

See fork(2) for more info on this syscall.

vfork

See vfork(2) for more info on this syscall.

execve

See execve(2) for more info on this syscall.

exit

See exit(2) for more info on this syscall.

wait4

See wait4(2) for more info on this syscall.

kill

See kill(2) for more info on this syscall.

uname

See uname(2) for more info on this syscall.

semget

See semget(2) for more info on this syscall.

semop

See semop(2) for more info on this syscall.

semctl

See semctl(2) for more info on this syscall.

shmdt

See shmdt(2) for more info on this syscall.

msgget

See msgget(2) for more info on this syscall.

msgsnd

See msgsnd(2) for more info on this syscall.

msgrcv

See msgrcv(2) for more info on this syscall.

msgctl

See msgctl(2) for more info on this syscall.

fcntl

See fcntl(2) for more info on this syscall.

flock

See flock(2) for more info on this syscall.

fsync

See fsync(2) for more info on this syscall.

fdatasync

See fdatasync(2) for more info on this syscall.

truncate

See truncate(2) for more info on this syscall.

ftruncate

See ftruncate(2) for more info on this syscall.

getdents

See getdents(2) for more info on this syscall.

getcwd

See getcwd(2) for more info on this syscall.

chdir

See chdir(2) for more info on this syscall.

fchdir

See fchdir(2) for more info on this syscall.

rename

See rename(2) for more info on this syscall.

mkdir

See mkdir(2) for more info on this syscall.

rmdir

See rmdir(2) for more info on this syscall.

creat

See creat(2) for more info on this syscall.

link

See link(2) for more info on this syscall.

unlink

See unlink(2) for more info on this syscall.

symlink

See symlink(2) for more info on this syscall.

readlink

See readlink(2) for more info on this syscall.

chmod

See chmod(2) for more info on this syscall.

fchmod

See fchmod(2) for more info on this syscall.

chown

See chown(2) for more info on this syscall.

fchown

See fchown(2) for more info on this syscall.

lchown

See lchown(2) for more info on this syscall.

umask

See umask(2) for more info on this syscall.

gettimeofday

See gettimeofday(2) for more info on this syscall.

getrlimit

See getrlimit(2) for more info on this syscall.

getrusage

See getrusage(2) for more info on this syscall.

sysinfo

See sysinfo(2) for more info on this syscall.

times

See times(2) for more info on this syscall.

ptrace

See ptrace(2) for more info on this syscall.

getuid

See getuid(2) for more info on this syscall.

syslog

See syslog(2) for more info on this syscall.

getgid

See getgid(2) for more info on this syscall.

setuid

See setuid(2) for more info on this syscall.

setgid

See setgid(2) for more info on this syscall.

geteuid

See geteuid(2) for more info on this syscall.

getegid

See getegid(2) for more info on this syscall.

setpgid

See setpgid(2) for more info on this syscall.

getppid

See getppid(2) for more info on this syscall.

getpgrp

See getpgrp(2) for more info on this syscall.

setsid

See setsid(2) for more info on this syscall.

setreuid

See setreuid(2) for more info on this syscall.

setregid

See setregid(2) for more info on this syscall.

getgroups

See getgroups(2) for more info on this syscall.

setgroups

See setgroups(2) for more info on this syscall.

setresuid

See setresuid(2) for more info on this syscall.

getresuid

See getresuid(2) for more info on this syscall.

setresgid

See setresgid(2) for more info on this syscall.

getresgid

See getresgid(2) for more info on this syscall.

getpgid

See getpgid(2) for more info on this syscall.

setfsuid

See setfsuid(2) for more info on this syscall.

setfsgid

See setfsgid(2) for more info on this syscall.

getsid

See getsid(2) for more info on this syscall.

capget

See capget(2) for more info on this syscall.

capset

See capset(2) for more info on this syscall.

rt_sigpending

See rt_sigpending(2) for more info on this syscall.

rt_sigtimedwait

See rt_sigtimedwait(2) for more info on this syscall.

rt_sigqueueinfo

See rt_sigqueueinfo(2) for more info on this syscall.

rt_sigsuspend

See rt_sigsuspend(2) for more info on this syscall.

sigaltstack

See sigaltstack(2) for more info on this syscall.

utime

See utime(2) for more info on this syscall.

mknod

See mknod(2) for more info on this syscall.

uselib

NOTE: uselib is not implemented in the kernel.

personality

See personality(2) for more info on this syscall.

ustat

See ustat(2) for more info on this syscall.

statfs

See statfs(2) for more info on this syscall.

fstatfs

See fstatfs(2) for more info on this syscall.

sysfs

See sysfs(2) for more info on this syscall.

getpriority

See getpriority(2) for more info on this syscall.

setpriority

See setpriority(2) for more info on this syscall.

sched_setparam

See sched_setparam(2) for more info on this syscall.

sched_getparam

See sched_getparam(2) for more info on this syscall.

sched_setscheduler

See sched_setscheduler(2) for more info on this syscall.

sched_getscheduler

See sched_getscheduler(2) for more info on this syscall.

sched_get_priority_max

See sched_get_priority_max(2) for more info on this syscall.

sched_get_priority_min

See sched_get_priority_min(2) for more info on this syscall.

sched_rr_get_interval

See sched_rr_get_interval(2) for more info on this syscall.

mlock

See mlock(2) for more info on this syscall.

munlock

See munlock(2) for more info on this syscall.

mlockall

See mlockall(2) for more info on this syscall.

munlockall

See munlockall(2) for more info on this syscall.

vhangup

See vhangup(2) for more info on this syscall.

modify_ldt

See modify_ldt(2) for more info on this syscall.

pivot_root

See pivot_root(2) for more info on this syscall.

_sysctl

See _sysctl(2) for more info on this syscall.

prctl

See prctl(2) for more info on this syscall.

arch_prctl

See arch_prctl(2) for more info on this syscall.

adjtimex

See adjtimex(2) for more info on this syscall.

setrlimit

See setrlimit(2) for more info on this syscall.

chroot

See chroot(2) for more info on this syscall.

sync

See sync(2) for more info on this syscall.

acct

See acct(2) for more info on this syscall.

settimeofday

See settimeofday(2) for more info on this syscall.

mount

See mount(2) for more info on this syscall.

umount2

See umount2(2) for more info on this syscall.

swapon

See swapon(2) for more info on this syscall.

swapoff

See swapoff(2) for more info on this syscall.

reboot

See reboot(2) for more info on this syscall.

sethostname

See sethostname(2) for more info on this syscall.

setdomainname

See setdomainname(2) for more info on this syscall.

iopl

See iopl(2) for more info on this syscall.

ioperm

See ioperm(2) for more info on this syscall.

create_module

NOTE: create_module is not implemented in the kernel.

init_module

See init_module(2) for more info on this syscall.

delete_module

See delete_module(2) for more info on this syscall.

get_kernel_syms

NOTE: get_kernel_syms is not implemented in the kernel.

query_module

NOTE: query_module is not implemented in the kernel.

quotactl

See quotactl(2) for more info on this syscall.

nfsservctl

NOTE: nfsservctl is not implemented in the kernel.

getpmsg

NOTE: getpmsg is not implemented in the kernel.

putpmsg

NOTE: putpmsg is not implemented in the kernel.

afs_syscall

NOTE: afs_syscall is not implemented in the kernel.

tuxcall

NOTE: tuxcall is not implemented in the kernel.

security

NOTE: security is not implemented in the kernel.

gettid

See gettid(2) for more info on this syscall.

readahead

See readahead(2) for more info on this syscall.

setxattr

See setxattr(2) for more info on this syscall.

lsetxattr

See lsetxattr(2) for more info on this syscall.

fsetxattr

See fsetxattr(2) for more info on this syscall.

getxattr

See getxattr(2) for more info on this syscall.

lgetxattr

See lgetxattr(2) for more info on this syscall.

fgetxattr

See fgetxattr(2) for more info on this syscall.

listxattr

See listxattr(2) for more info on this syscall.

llistxattr

See llistxattr(2) for more info on this syscall.

flistxattr

See flistxattr(2) for more info on this syscall.

removexattr

See removexattr(2) for more info on this syscall.

lremovexattr

See lremovexattr(2) for more info on this syscall.

fremovexattr

See fremovexattr(2) for more info on this syscall.

tkill

See tkill(2) for more info on this syscall.

time

See time(2) for more info on this syscall.

futex

See futex(2) for more info on this syscall.

sched_setaffinity

See sched_setaffinity(2) for more info on this syscall.

sched_getaffinity

See sched_getaffinity(2) for more info on this syscall.

set_thread_area

NOTE: set_thread_area is not implemented in the kernel.

io_setup

See io_setup(2) for more info on this syscall.

io_destroy

See io_destroy(2) for more info on this syscall.

io_getevents

See io_getevents(2) for more info on this syscall.

io_submit

See io_submit(2) for more info on this syscall.

io_cancel

See io_cancel(2) for more info on this syscall.

get_thread_area

NOTE: get_thread_area is not implemented in the kernel.

lookup_dcookie

See lookup_dcookie(2) for more info on this syscall.

epoll_create

See epoll_create(2) for more info on this syscall.

epoll_ctl_old

NOTE: epoll_ctl_old is not implemented in the kernel.

epoll_wait_old

NOTE: epoll_wait_old is not implemented in the kernel.

remap_file_pages

See remap_file_pages(2) for more info on this syscall.

getdents64

See getdents64(2) for more info on this syscall.

set_tid_address

See set_tid_address(2) for more info on this syscall.

restart_syscall

See restart_syscall(2) for more info on this syscall.

semtimedop

See semtimedop(2) for more info on this syscall.

fadvise64

See fadvise64(2) for more info on this syscall.

timer_create

See timer_create(2) for more info on this syscall.

timer_settime

See timer_settime(2) for more info on this syscall.

timer_gettime

See timer_gettime(2) for more info on this syscall.

timer_getoverrun

See timer_getoverrun(2) for more info on this syscall.

timer_delete

See timer_delete(2) for more info on this syscall.

clock_settime

See clock_settime(2) for more info on this syscall.

clock_gettime

See clock_gettime(2) for more info on this syscall.

clock_getres

See clock_getres(2) for more info on this syscall.

clock_nanosleep

See clock_nanosleep(2) for more info on this syscall.

exit_group

See exit_group(2) for more info on this syscall.

epoll_wait

See epoll_wait(2) for more info on this syscall.

epoll_ctl

See epoll_ctl(2) for more info on this syscall.

tgkill

See tgkill(2) for more info on this syscall.

utimes

See utimes(2) for more info on this syscall.

vserver

NOTE: vserver is not implemented in the kernel.

mbind

See mbind(2) for more info on this syscall.

set_mempolicy

See set_mempolicy(2) for more info on this syscall.

get_mempolicy

See get_mempolicy(2) for more info on this syscall.

mq_open

See mq_open(2) for more info on this syscall.

mq_unlink

See mq_unlink(2) for more info on this syscall.

mq_timedsend

See mq_timedsend(2) for more info on this syscall.

mq_timedreceive

See mq_timedreceive(2) for more info on this syscall.

mq_notify

See mq_notify(2) for more info on this syscall.

mq_getsetattr

See mq_getsetattr(2) for more info on this syscall.

kexec_load

See kexec_load(2) for more info on this syscall.

waitid

See waitid(2) for more info on this syscall.

add_key

See add_key(2) for more info on this syscall.

request_key

See request_key(2) for more info on this syscall.

keyctl

See keyctl(2) for more info on this syscall.

ioprio_set

See ioprio_set(2) for more info on this syscall.

ioprio_get

See ioprio_get(2) for more info on this syscall.

inotify_init

See inotify_init(2) for more info on this syscall.

inotify_add_watch

See inotify_add_watch(2) for more info on this syscall.

inotify_rm_watch

See inotify_rm_watch(2) for more info on this syscall.

migrate_pages

See migrate_pages(2) for more info on this syscall.

openat

See openat(2) for more info on this syscall.

mkdirat

See mkdirat(2) for more info on this syscall.

mknodat

See mknodat(2) for more info on this syscall.

fchownat

See fchownat(2) for more info on this syscall.

futimesat

See futimesat(2) for more info on this syscall.

newfstatat

See newfstatat(2) for more info on this syscall.

unlinkat

See unlinkat(2) for more info on this syscall.

renameat

See renameat(2) for more info on this syscall.

linkat

See linkat(2) for more info on this syscall.

symlinkat

See symlinkat(2) for more info on this syscall.

readlinkat

See readlinkat(2) for more info on this syscall.

fchmodat

See fchmodat(2) for more info on this syscall.

faccessat

See faccessat(2) for more info on this syscall.

pselect6

See pselect6(2) for more info on this syscall.

ppoll

See ppoll(2) for more info on this syscall.

unshare

See unshare(2) for more info on this syscall.

set_robust_list

See set_robust_list(2) for more info on this syscall.

get_robust_list

See get_robust_list(2) for more info on this syscall.

splice

See splice(2) for more info on this syscall.

tee

See tee(2) for more info on this syscall.

sync_file_range

See sync_file_range(2) for more info on this syscall.

vmsplice

See vmsplice(2) for more info on this syscall.

move_pages

See move_pages(2) for more info on this syscall.

utimensat

See utimensat(2) for more info on this syscall.

epoll_pwait

See epoll_pwait(2) for more info on this syscall.

signalfd

See signalfd(2) for more info on this syscall.

timerfd_create

See timerfd_create(2) for more info on this syscall.

eventfd

See eventfd(2) for more info on this syscall.

fallocate

See fallocate(2) for more info on this syscall.

timerfd_settime

See timerfd_settime(2) for more info on this syscall.

timerfd_gettime

See timerfd_gettime(2) for more info on this syscall.

accept4

See accept4(2) for more info on this syscall.

signalfd4

See signalfd4(2) for more info on this syscall.

eventfd2

See eventfd2(2) for more info on this syscall.

epoll_create1

See epoll_create1(2) for more info on this syscall.

dup3

See dup3(2) for more info on this syscall.

pipe2

See pipe2(2) for more info on this syscall.

inotify_init1

See inotify_init1(2) for more info on this syscall.

preadv

See preadv(2) for more info on this syscall.

pwritev

See pwritev(2) for more info on this syscall.

rt_tgsigqueueinfo

See rt_tgsigqueueinfo(2) for more info on this syscall.

perf_event_open

See perf_event_open(2) for more info on this syscall.

recvmmsg

See recvmmsg(2) for more info on this syscall.

fanotify_init

See fanotify_init(2) for more info on this syscall.

fanotify_mark

See fanotify_mark(2) for more info on this syscall.

prlimit64

See prlimit64(2) for more info on this syscall.

name_to_handle_at

See name_to_handle_at(2) for more info on this syscall.

open_by_handle_at

See open_by_handle_at(2) for more info on this syscall.

clock_adjtime

See clock_adjtime(2) for more info on this syscall.

syncfs

See syncfs(2) for more info on this syscall.

sendmmsg

See sendmmsg(2) for more info on this syscall.

setns

See setns(2) for more info on this syscall.

getcpu

See getcpu(2) for more info on this syscall.

process_vm_readv

See process_vm_readv(2) for more info on this syscall.

process_vm_writev

See process_vm_writev(2) for more info on this syscall.

kcmp

See kcmp(2) for more info on this syscall.

finit_module

See finit_module(2) for more info on this syscall.

sched_setattr

See sched_setattr(2) for more info on this syscall.

sched_getattr

See sched_getattr(2) for more info on this syscall.

renameat2

See renameat2(2) for more info on this syscall.

seccomp

See seccomp(2) for more info on this syscall.

getrandom

See getrandom(2) for more info on this syscall.

memfd_create

See memfd_create(2) for more info on this syscall.

kexec_file_load

See kexec_file_load(2) for more info on this syscall.

bpf

See bpf(2) for more info on this syscall.

execveat

See execveat(2) for more info on this syscall.

userfaultfd

See userfaultfd(2) for more info on this syscall.

membarrier

See membarrier(2) for more info on this syscall.

mlock2

See mlock2(2) for more info on this syscall.

copy_file_range

See copy_file_range(2) for more info on this syscall.

preadv2

See preadv2(2) for more info on this syscall.

pwritev2

See pwritev2(2) for more info on this syscall.

pkey_mprotect

See pkey_mprotect(2) for more info on this syscall.

pkey_alloc

See pkey_alloc(2) for more info on this syscall.

pkey_free

See pkey_free(2) for more info on this syscall.

statx

See statx(2) for more info on this syscall.

io_pgetevents

See io_pgetevents(2) for more info on this syscall.

rseq

See rseq(2) for more info on this syscall.

pidfd_send_signal

See pidfd_send_signal(2) for more info on this syscall.

io_uring_setup

See io_uring_setup(2) for more info on this syscall.

io_uring_enter

See io_uring_enter(2) for more info on this syscall.

io_uring_register

See io_uring_register(2) for more info on this syscall.

open_tree

See open_tree(2) for more info on this syscall.

move_mount

See move_mount(2) for more info on this syscall.

fsopen

See fsopen(2) for more info on this syscall.

fsconfig

See fsconfig(2) for more info on this syscall.

fsmount

See fsmount(2) for more info on this syscall.

fspick

See fspick(2) for more info on this syscall.

pidfd_open

See pidfd_open(2) for more info on this syscall.

clone3

See clone3(2) for more info on this syscall.

close_range

See close_range(2) for more info on this syscall.

openat2

See openat2(2) for more info on this syscall.

pidfd_getfd

See pidfd_getfd(2) for more info on this syscall.

faccessat2

See faccessat2(2) for more info on this syscall.

process_madvise

See process_madvise(2) for more info on this syscall.

epoll_pwait2

See epoll_pwait2(2) for more info on this syscall.

mount_setattr

See mount_setattr(2) for more info on this syscall.

quotactl_fd

See quotactl_fd(2) for more info on this syscall.

landlock_create_ruleset

See landlock_create_ruleset(2) for more info on this syscall.

landlock_add_rule

See landlock_add_rule(2) for more info on this syscall.

landlock_restrict_self

See landlock_restrict_self(2) for more info on this syscall.

memfd_secret

See memfd_secret(2) for more info on this syscall.

process_mrelease

See process_mrelease(2) for more info on this syscall.

futex_waitv

See futex_waitv(2) for more info on this syscall.

set_mempolicy_home_node

See set_mempolicy_home_node(2) for more info on this syscall.

Implementations

impl Sysno

pub const fn new(id: usize) -> Option<Self>

Constructs a new syscall from the given ID. If the ID does not represent a valid syscall, returns None.

pub const fn name(&self) -> &'static str

Returns the name of the syscall.

pub const fn next(&self) -> Option<Self>

Returns the next syscall in the table. Returns None if this is the last syscall.

pub const fn first() -> Self

Returns the first syscall in the table.

pub const fn last() -> Self

Returns the last syscall in the table.

pub const fn id(&self) -> i32

Returns the syscall number.

pub const fn len() -> usize

👎Deprecated:
Sysno::len() is misleading. Use Sysno::table_size() instead.

Returns the length of the syscall table, including any gaps.

pub const fn count() -> usize

Returns the total number of valid syscalls.

pub const fn table_size() -> usize

Returns the length of the syscall table, including any gaps. This is not the same thing as the total number of syscalls.

pub fn iter() -> impl Iterator<Item = Self>

Returns an iterator that iterates over all possible syscalls.