Struct syscalls::Errno

pub struct Errno(_);

Implementations

impl Errno

pub const EPERM: Errno = _

Operation not permitted

pub const ENOENT: Errno = _

No such file or directory

pub const ESRCH: Errno = _

No such process

pub const EINTR: Errno = _

Interrupted system call

pub const EIO: Errno = _

I/O error

pub const ENXIO: Errno = _

No such device or address

pub const E2BIG: Errno = _

Argument list too long

pub const ENOEXEC: Errno = _

Exec format error

pub const EBADF: Errno = _

Bad file number

pub const ECHILD: Errno = _

No child processes

pub const EAGAIN: Errno = _

Try again

pub const ENOMEM: Errno = _

Out of memory

pub const EACCES: Errno = _

Permission denied

pub const EFAULT: Errno = _

Bad address

pub const ENOTBLK: Errno = _

Block device required

pub const EBUSY: Errno = _

Device or resource busy

pub const EEXIST: Errno = _

File exists

pub const EXDEV: Errno = _

Cross-device link

pub const ENODEV: Errno = _

No such device

pub const ENOTDIR: Errno = _

Not a directory

pub const EISDIR: Errno = _

Is a directory

pub const EINVAL: Errno = _

Invalid argument

pub const ENFILE: Errno = _

File table overflow

pub const EMFILE: Errno = _

Too many open files

pub const ENOTTY: Errno = _

Not a typewriter

pub const ETXTBSY: Errno = _

Text file busy

pub const EFBIG: Errno = _

File too large

pub const ENOSPC: Errno = _

No space left on device

pub const ESPIPE: Errno = _

Illegal seek

pub const EROFS: Errno = _

Read-only file system

pub const EMLINK: Errno = _

Too many links

pub const EPIPE: Errno = _

Broken pipe

pub const EDOM: Errno = _

Math argument out of domain of func

pub const ERANGE: Errno = _

Math result not representable

pub const EDEADLK: Errno = _

Resource deadlock would occur

pub const ENAMETOOLONG: Errno = _

File name too long

pub const ENOLCK: Errno = _

No record locks available

pub const ENOSYS: Errno = _

Invalid system call number

pub const ENOTEMPTY: Errno = _

Directory not empty

pub const ELOOP: Errno = _

Too many symbolic links encountered

pub const ENOMSG: Errno = _

No message of desired type

pub const EIDRM: Errno = _

Identifier removed

pub const ECHRNG: Errno = _

Channel number out of range

pub const EL2NSYNC: Errno = _

Level 2 not synchronized

pub const EL3HLT: Errno = _

Level 3 halted

pub const EL3RST: Errno = _

Level 3 reset

pub const ELNRNG: Errno = _

Link number out of range

pub const EUNATCH: Errno = _

Protocol driver not attached

pub const ENOCSI: Errno = _

No CSI structure available

pub const EL2HLT: Errno = _

Level 2 halted

pub const EBADE: Errno = _

Invalid exchange

pub const EBADR: Errno = _

Invalid request descriptor

pub const EXFULL: Errno = _

Exchange full

pub const ENOANO: Errno = _

No anode

pub const EBADRQC: Errno = _

Invalid request code

pub const EBADSLT: Errno = _

Invalid slot

pub const EBFONT: Errno = _

Bad font file format

pub const ENOSTR: Errno = _

Device not a stream

pub const ENODATA: Errno = _

No data available

pub const ETIME: Errno = _

Timer expired

pub const ENOSR: Errno = _

Out of streams resources

pub const ENONET: Errno = _

Machine is not on the network

pub const ENOPKG: Errno = _

Package not installed

pub const EREMOTE: Errno = _

Object is remote

pub const ENOLINK: Errno = _

Link has been severed

pub const EADV: Errno = _

Advertise error

pub const ESRMNT: Errno = _

Srmount error

pub const ECOMM: Errno = _

Communication error on send

pub const EPROTO: Errno = _

Protocol error

pub const EMULTIHOP: Errno = _

Multihop attempted

pub const EDOTDOT: Errno = _

RFS specific error

pub const EBADMSG: Errno = _

Not a data message

pub const EOVERFLOW: Errno = _

Value too large for defined data type

pub const ENOTUNIQ: Errno = _

Name not unique on network

pub const EBADFD: Errno = _

File descriptor in bad state

pub const EREMCHG: Errno = _

Remote address changed

pub const ELIBACC: Errno = _

Can not access a needed shared library

pub const ELIBBAD: Errno = _

Accessing a corrupted shared library

pub const ELIBSCN: Errno = _

.lib section in a.out corrupted

pub const ELIBMAX: Errno = _

Attempting to link in too many shared libraries

pub const ELIBEXEC: Errno = _

Cannot exec a shared library directly

pub const EILSEQ: Errno = _

Illegal byte sequence

pub const ERESTART: Errno = _

Interrupted system call should be restarted

pub const ESTRPIPE: Errno = _

Streams pipe error

pub const EUSERS: Errno = _

Too many users

pub const ENOTSOCK: Errno = _

Socket operation on non-socket

pub const EDESTADDRREQ: Errno = _

Destination address required

pub const EMSGSIZE: Errno = _

Message too long

pub const EPROTOTYPE: Errno = _

Protocol wrong type for socket

pub const ENOPROTOOPT: Errno = _

Protocol not available

pub const EPROTONOSUPPORT: Errno = _

Protocol not supported

pub const ESOCKTNOSUPPORT: Errno = _

Socket type not supported

pub const EOPNOTSUPP: Errno = _

Operation not supported on transport endpoint

pub const EPFNOSUPPORT: Errno = _

Protocol family not supported

pub const EAFNOSUPPORT: Errno = _

Address family not supported by protocol

pub const EADDRINUSE: Errno = _

Address already in use

pub const EADDRNOTAVAIL: Errno = _

Cannot assign requested address

pub const ENETDOWN: Errno = _

Network is down

pub const ENETUNREACH: Errno = _

Network is unreachable

pub const ENETRESET: Errno = _

Network dropped connection because of reset

pub const ECONNABORTED: Errno = _

Software caused connection abort

pub const ECONNRESET: Errno = _

Connection reset by peer

pub const ENOBUFS: Errno = _

No buffer space available

pub const EISCONN: Errno = _

Transport endpoint is already connected

pub const ENOTCONN: Errno = _

Transport endpoint is not connected

pub const ESHUTDOWN: Errno = _

Cannot send after transport endpoint shutdown

pub const ETOOMANYREFS: Errno = _

Too many references: cannot splice

pub const ETIMEDOUT: Errno = _

Connection timed out

pub const ECONNREFUSED: Errno = _

Connection refused

pub const EHOSTDOWN: Errno = _

Host is down

pub const EHOSTUNREACH: Errno = _

No route to host

pub const EALREADY: Errno = _

Operation already in progress

pub const EINPROGRESS: Errno = _

Operation now in progress

pub const ESTALE: Errno = _

Stale file handle

pub const EUCLEAN: Errno = _

Structure needs cleaning

pub const ENOTNAM: Errno = _

Not a XENIX named type file

pub const ENAVAIL: Errno = _

No XENIX semaphores available

pub const EISNAM: Errno = _

Is a named type file

pub const EREMOTEIO: Errno = _

Remote I/O error

pub const EDQUOT: Errno = _

Quota exceeded

pub const ENOMEDIUM: Errno = _

No medium found

pub const EMEDIUMTYPE: Errno = _

Wrong medium type

pub const ECANCELED: Errno = _

Operation Canceled

pub const ENOKEY: Errno = _

Required key not available

pub const EKEYEXPIRED: Errno = _

Key has expired

pub const EKEYREVOKED: Errno = _

Key has been revoked

pub const EKEYREJECTED: Errno = _

Key was rejected by service

pub const EOWNERDEAD: Errno = _

Owner died

pub const ENOTRECOVERABLE: Errno = _

State not recoverable

pub const ERFKILL: Errno = _

Operation not possible due to RF-kill

pub const EHWPOISON: Errno = _

Memory page has hardware error

pub const ERESTARTSYS: Errno = _

Restart syscall

pub const ERESTARTNOINTR: Errno = _

Restart if no interrupt

pub const ERESTARTNOHAND: Errno = _

restart if no handler..

pub const ENOIOCTLCMD: Errno = _

No ioctl command

pub const ERESTART_RESTARTBLOCK: Errno = _

restart by calling sys_restart_syscall

pub const EPROBE_DEFER: Errno = _

Driver requests probe retry

pub const EOPENSTALE: Errno = _

open found a stale dentry

pub const ENOPARAM: Errno = _

Parameter not supported

pub const EBADHANDLE: Errno = _

Illegal NFS file handle

pub const ENOTSYNC: Errno = _

Update synchronization mismatch

pub const EBADCOOKIE: Errno = _

Cookie is stale

pub const ENOTSUPP: Errno = _

Operation is not supported

pub const ETOOSMALL: Errno = _

Buffer or request is too small

pub const ESERVERFAULT: Errno = _

An untranslatable error occurred

pub const EBADTYPE: Errno = _

Type not supported by server

pub const EJUKEBOX: Errno = _

Request initiated, but will not complete before timeout

pub const EIOCBQUEUED: Errno = _

iocb queued, will get completion event

pub const ERECALLCONFLICT: Errno = _

conflict with recalled state

pub const ENOGRACE: Errno = _

NFS file lock reclaim refused

pub fn name_and_description(&self) -> Option<(&'static str, &'static str)>

Returns a pair containing the name of the error and a string describing the error.

impl Errno

pub const EWOULDBLOCK: Self = Self::EAGAIN

Operation would block. This is the same as Errno::EAGAIN.

pub const EDEADLOCK: Self = Self::EDEADLK

Same as Errno::EDEADLK.

pub fn new(num: i32) -> Self

Creates a new Errno.

pub fn into_raw(self) -> i32

Converts the Errno into a raw i32.

pub fn is_valid(&self) -> bool

Returns true if the error code is valid (i.e., less than 4096).

pub fn from_ret(value: usize) -> Result<usize, Errno>

Converts a raw syscall return value to a result.

pub fn last() -> Self

Returns the last error that occurred.

pub fn result<T>(value: T) -> Result<T, Errno>where
T: ErrnoSentinel + PartialEq<T>,

Converts a value into an Errno.

pub fn name(&self) -> Option<&'static str>

Returns the name of the error. If the internal error code is unknown or invalid, None is returned.

pub fn description(&self) -> Option<&'static str>

Returns the error description. If the internal error code is unknown or invalid, None is returned.

pub fn from_io_error(err: Error) -> Option<Self>

Converts an std::io::Error into an Errno if possible. Since an error code is just one of the few possible error types that std::io::Error can represent, this will return None if the conversion is not possible.

A From<std::io::Error> implementation is not provided because this conversion can fail. However, the reverse is possible, so that is provided as a From implementation.