syscalls

Struct Errno

pub struct Errno(/* private fields */);

Implementations

impl Errno

pub const EPERM: Errno

Operation not permitted

pub const ENOENT: Errno

No such file or directory

pub const ESRCH: Errno

No such process

pub const EINTR: Errno

Interrupted system call

pub const EIO: Errno

I/O error

pub const ENXIO: Errno

No such device or address

pub const E2BIG: Errno

Argument list too long

pub const ENOEXEC: Errno

Exec format error

pub const EBADF: Errno

Bad file number

pub const ECHILD: Errno

No child processes

pub const EAGAIN: Errno

Try again

pub const ENOMEM: Errno

Out of memory

pub const EACCES: Errno

Permission denied

pub const EFAULT: Errno

Bad address

pub const ENOTBLK: Errno

Block device required

pub const EBUSY: Errno

Device or resource busy

pub const EEXIST: Errno

File exists

pub const EXDEV: Errno

Cross-device link

pub const ENODEV: Errno

No such device

pub const ENOTDIR: Errno

Not a directory

pub const EISDIR: Errno

Is a directory

pub const EINVAL: Errno

Invalid argument

pub const ENFILE: Errno

File table overflow

pub const EMFILE: Errno

Too many open files

pub const ENOTTY: Errno

Not a typewriter

pub const ETXTBSY: Errno

Text file busy

pub const EFBIG: Errno

File too large

pub const ENOSPC: Errno

No space left on device

pub const ESPIPE: Errno

Illegal seek

pub const EROFS: Errno

Read-only file system

pub const EMLINK: Errno

Too many links

pub const EPIPE: Errno

Broken pipe

pub const EDOM: Errno

Math argument out of domain of func

pub const ERANGE: Errno

Math result not representable

pub const EDEADLK: Errno

Resource deadlock would occur

pub const ENAMETOOLONG: Errno

File name too long

pub const ENOLCK: Errno

No record locks available

pub const ENOSYS: Errno

Invalid system call number

pub const ENOTEMPTY: Errno

Directory not empty

pub const ELOOP: Errno

Too many symbolic links encountered

pub const ENOMSG: Errno

No message of desired type

pub const EIDRM: Errno

Identifier removed

pub const ECHRNG: Errno

Channel number out of range

pub const EL2NSYNC: Errno

Level 2 not synchronized

pub const EL3HLT: Errno

Level 3 halted

pub const EL3RST: Errno

Level 3 reset

pub const ELNRNG: Errno

Link number out of range

pub const EUNATCH: Errno

Protocol driver not attached

pub const ENOCSI: Errno

No CSI structure available

pub const EL2HLT: Errno

Level 2 halted

pub const EBADE: Errno

Invalid exchange

pub const EBADR: Errno

Invalid request descriptor

pub const EXFULL: Errno

Exchange full

pub const ENOANO: Errno

No anode

pub const EBADRQC: Errno

Invalid request code

pub const EBADSLT: Errno

Invalid slot

pub const EBFONT: Errno

Bad font file format

pub const ENOSTR: Errno

Device not a stream

pub const ENODATA: Errno

No data available

pub const ETIME: Errno

Timer expired

pub const ENOSR: Errno

Out of streams resources

pub const ENONET: Errno

Machine is not on the network

pub const ENOPKG: Errno

Package not installed

pub const EREMOTE: Errno

Object is remote

pub const ENOLINK: Errno

Link has been severed

pub const EADV: Errno

Advertise error

pub const ESRMNT: Errno

Srmount error

pub const ECOMM: Errno

Communication error on send

pub const EPROTO: Errno

Protocol error

pub const EMULTIHOP: Errno

Multihop attempted

pub const EDOTDOT: Errno

RFS specific error

pub const EBADMSG: Errno

Not a data message

pub const EOVERFLOW: Errno

Value too large for defined data type

pub const ENOTUNIQ: Errno

Name not unique on network

pub const EBADFD: Errno

File descriptor in bad state

pub const EREMCHG: Errno

Remote address changed

pub const ELIBACC: Errno

Can not access a needed shared library

pub const ELIBBAD: Errno

Accessing a corrupted shared library

pub const ELIBSCN: Errno

.lib section in a.out corrupted

pub const ELIBMAX: Errno

Attempting to link in too many shared libraries

pub const ELIBEXEC: Errno

Cannot exec a shared library directly

pub const EILSEQ: Errno

Illegal byte sequence

pub const ERESTART: Errno

Interrupted system call should be restarted

pub const ESTRPIPE: Errno

Streams pipe error

pub const EUSERS: Errno

Too many users

pub const ENOTSOCK: Errno

Socket operation on non-socket

pub const EDESTADDRREQ: Errno

Destination address required

pub const EMSGSIZE: Errno

Message too long

pub const EPROTOTYPE: Errno

Protocol wrong type for socket

pub const ENOPROTOOPT: Errno

Protocol not available

pub const EPROTONOSUPPORT: Errno

Protocol not supported

pub const ESOCKTNOSUPPORT: Errno

Socket type not supported

pub const EOPNOTSUPP: Errno

Operation not supported on transport endpoint

pub const EPFNOSUPPORT: Errno

Protocol family not supported

pub const EAFNOSUPPORT: Errno

Address family not supported by protocol

pub const EADDRINUSE: Errno

Address already in use

pub const EADDRNOTAVAIL: Errno

Cannot assign requested address

pub const ENETDOWN: Errno

Network is down

pub const ENETUNREACH: Errno

Network is unreachable

pub const ENETRESET: Errno

Network dropped connection because of reset

pub const ECONNABORTED: Errno

Software caused connection abort

pub const ECONNRESET: Errno

Connection reset by peer

pub const ENOBUFS: Errno

No buffer space available

pub const EISCONN: Errno

Transport endpoint is already connected

pub const ENOTCONN: Errno

Transport endpoint is not connected

pub const ESHUTDOWN: Errno

Cannot send after transport endpoint shutdown

pub const ETOOMANYREFS: Errno

Too many references: cannot splice

pub const ETIMEDOUT: Errno

Connection timed out

pub const ECONNREFUSED: Errno

Connection refused

pub const EHOSTDOWN: Errno

Host is down

pub const EHOSTUNREACH: Errno

No route to host

pub const EALREADY: Errno

Operation already in progress

pub const EINPROGRESS: Errno

Operation now in progress

pub const ESTALE: Errno

Stale file handle

pub const EUCLEAN: Errno

Structure needs cleaning

pub const ENOTNAM: Errno

Not a XENIX named type file

pub const ENAVAIL: Errno

No XENIX semaphores available

pub const EISNAM: Errno

Is a named type file

pub const EREMOTEIO: Errno

Remote I/O error

pub const EDQUOT: Errno

Quota exceeded

pub const ENOMEDIUM: Errno

No medium found

pub const EMEDIUMTYPE: Errno

Wrong medium type

pub const ECANCELED: Errno

Operation Canceled

pub const ENOKEY: Errno

Required key not available

pub const EKEYEXPIRED: Errno

Key has expired

pub const EKEYREVOKED: Errno

Key has been revoked

pub const EKEYREJECTED: Errno

Key was rejected by service

pub const EOWNERDEAD: Errno

Owner died

pub const ENOTRECOVERABLE: Errno

State not recoverable

pub const ERFKILL: Errno

Operation not possible due to RF-kill

pub const EHWPOISON: Errno

Memory page has hardware error

pub const ERESTARTSYS: Errno

Restart syscall

pub const ERESTARTNOINTR: Errno

Restart if no interrupt

pub const ERESTARTNOHAND: Errno

restart if no handler..

pub const ENOIOCTLCMD: Errno

No ioctl command

pub const ERESTART_RESTARTBLOCK: Errno

restart by calling sys_restart_syscall

pub const EPROBE_DEFER: Errno

Driver requests probe retry

pub const EOPENSTALE: Errno

open found a stale dentry

pub const ENOPARAM: Errno

Parameter not supported

pub const EBADHANDLE: Errno

Illegal NFS file handle

pub const ENOTSYNC: Errno

Update synchronization mismatch

pub const EBADCOOKIE: Errno

Cookie is stale

pub const ENOTSUPP: Errno

Operation is not supported

pub const ETOOSMALL: Errno

Buffer or request is too small

pub const ESERVERFAULT: Errno

An untranslatable error occurred

pub const EBADTYPE: Errno

Type not supported by server

pub const EJUKEBOX: Errno

Request initiated, but will not complete before timeout

pub const EIOCBQUEUED: Errno

iocb queued, will get completion event

pub const ERECALLCONFLICT: Errno

conflict with recalled state

pub const ENOGRACE: Errno

NFS file lock reclaim refused

pub fn name_and_description(&self) -> Option<(&'static str, &'static str)>

Returns a pair containing the name of the error and a string describing the error.

impl Errno

pub const EWOULDBLOCK: Self = Self::EAGAIN

Operation would block. This is the same as Errno::EAGAIN.

pub const EDEADLOCK: Self = Self::EDEADLK

Same as Errno::EDEADLK.

pub fn new(num: i32) -> Self

Creates a new Errno.

pub fn into_raw(self) -> i32

Converts the Errno into a raw i32.

pub fn is_valid(&self) -> bool

Returns true if the error code is valid (i.e., less than 4096).

pub fn from_ret(value: usize) -> Result<usize, Errno>

Converts a raw syscall return value to a result.

pub fn name(&self) -> Option<&'static str>

Returns the name of the error. If the internal error code is unknown or invalid, None is returned.

pub fn description(&self) -> Option<&'static str>

Returns the error description. If the internal error code is unknown or invalid, None is returned.